✨ Centralize stops management
