+++ src/views/BaseView.vue
@@ -2,67 +2,30 @@
 import { onMounted, ref } from "vue";
 import { onBeforeRouteUpdate, useRoute } from "vue-router";
 import SettingsComp from "@/components/BaseSettings.vue";
+import { fetchStops } from "@/store/TBM";
+// May wait for settings & use loader instead
+import { fetchPaused, settings, updateStoredSettings } from "@/store/Storage";
 import {
-  type FullyDescribedStopArea,
-  type StopArea,
-  type StopPoint,
-  fetchStops,
-  fetchStopAreaDetails,
-} from "@/store/TBM";
+  addStop,
+  provideQuery,
+  queryUpdated,
+  selectedStops,
+  stops,
+  getWantedStops,
+} from "@/store/StopsManager";
 
 const showSettingsButton = ref<HTMLButtonElement | null>(null);
 const settingsComp = ref<InstanceType<typeof SettingsComp> | null>(null);
 
 const route = useRoute();
-let query = { ...route.query };
-let queryInternallyUpdated = false;
+provideQuery(route.query);
 
-onMounted(queryUpdated);
+fetchPaused();
+
+onMounted(() => queryUpdated(route));
 onBeforeRouteUpdate((to) => queryUpdated(to));
 
-const selectedStops = ref<FullyDescribedStopArea[]>([]);
-const excludedStopPoints = ref<[StopArea["id"], StopPoint["id"]][]>([]);
-
-async function queryUpdated(to = route) {
-  if (queryInternallyUpdated) {
-    queryInternallyUpdated = false;
-    return;
-  }
-  query = { ...to.query };
-
-  for (const k of Object.keys(to.query)
-    .filter((k) => !isNaN(parseInt(k)))
-    .sort((a, b) => parseInt(a) - parseInt(b))) {
-    const providenStop = to.query[k] as string;
-
-    if (selectedStops.value.find((s) => s.name === providenStop)) continue;
-    const found = (await fetchStops(providenStop)).find((s) => s.name === providenStop);
-    if (!found) {
-      delete query[k];
-      queryInternallyUpdated = true;
-      router.push({ query });
-      continue;
-    }
-    const fullyDescribedStop = await fetchStopAreaDetails(found);
-    if (!fullyDescribedStop) {
-      delete query[k];
-      queryInternallyUpdated = true;
-      router.push({ query });
-      continue;
-    }
-    selectedStops.value = [...selectedStops.value, fullyDescribedStop];
-  }
-
-  const providenStops = Object.keys(query).map((k) => query[k]);
-  selectedStops.value = selectedStops.value.filter((s) => providenStops.includes(s.name));
-
-  excludedStopPoints.value = deserializeExcludedStopPoints(
-    typeof query["eSP"] === "string" ? query["eSP"] : "",
-  );
-}
-
 const stopInput = ref<string>("");
-const stops = ref<StopArea[]>([]);
 
 async function refreshStops() {
   if (await addCurrentStop()) return;
@@ -71,146 +34,13 @@ async function refreshStops() {
 }
 
 async function addCurrentStop() {
-  const alreadySelected = selectedStops.value.find((s) => s.name === stopInput.value);
-  if (alreadySelected) {
-    if (excludedStopPoints.value.find(([sp]) => sp === alreadySelected.id)) {
-      excludedStopPoints.value = excludedStopPoints.value.filter(([sp]) => sp != alreadySelected.id);
-      query["eSP"] = serializeExcludedStopPoints(excludedStopPoints.value);
-      if (!query["eSP"].length) delete query["eSP"];
-      queryInternallyUpdated = true;
-      router.push({ query });
-      return (stopInput.value = "");
-    } else return (stopInput.value = ""); //display error
-  }
-  const found = stops.value.find((s) => s.name === stopInput.value);
-  if (!found) return false; // display error
-  const fullyDescribedStop = await fetchStopAreaDetails(found);
-  if (!fullyDescribedStop) return false; // display error
-
-  query[Object.keys(query).length + 1] = fullyDescribedStop.name;
-  queryInternallyUpdated = true;
-  router.push({ query });
-
-  selectedStops.value = [...selectedStops.value, fullyDescribedStop];
-
-  stopInput.value = "";
-  stops.value = [];
-  return true;
-}
-
-function removeStop(stopArea: FullyDescribedStopArea) {
-  selectedStops.value = selectedStops.value.filter((s) => s.id != stopArea.id);
-
-  let queryNeedUpdate = false;
-
-  if (excludedStopPoints.value.find(([stopAreaId]) => stopAreaId === stopArea.id)) {
-    queryNeedUpdate = true;
-    excludedStopPoints.value = excludedStopPoints.value.filter(([stopAreaId]) => stopAreaId !== stopArea.id);
-    query["eSP"] = serializeExcludedStopPoints(excludedStopPoints.value);
-    if (!query["eSP"].length) delete query["eSP"];
+  const r = await addStop(stopInput.value);
+  if (r >= 0) {
+    stopInput.value = "";
+    if (r > 1) stops.value = [];
   }
 
-  Object.keys(query).forEach((k) => {
-    if (query[k] === stopArea.name) {
-      queryNeedUpdate = true;
-      delete query[k];
-    }
-  });
-
-  if (queryNeedUpdate) {
-    queryInternallyUpdated = true;
-    router.push({ query });
-  }
-}
-
-function removeStopPoint(stopArea: FullyDescribedStopArea, stopPoint: StopPoint) {
-  if (
-    excludedStopPoints.value.find(
-      ([stopAreaId, stopPointId]) => stopArea.id === stopAreaId && stopPoint.id === stopPointId,
-    )
-  )
-    return;
-
-  excludedStopPoints.value.push([stopArea.id, stopPoint.id]);
-
-  query["eSP"] = serializeExcludedStopPoints(excludedStopPoints.value);
-  if (!query["eSP"].length) delete query["eSP"];
-  queryInternallyUpdated = true;
-  router.push({ query });
-
-  if (stopArea.details.stopPoints.every((s) => excludedStopPoints.value.find((esp) => s.id === esp[1])))
-    removeStop(stopArea);
-}
-
-function deserializeExcludedStopPoints(serializedExcludedStopPoints: string) {
-  const stopAreas = serializedExcludedStopPoints.split(",");
-  if (!stopAreas.length) return [];
-
-  const excludedStopPoints = [] as [StopArea["id"], StopPoint["id"]][];
-
-  for (const stopAreaString of stopAreas) {
-    const stopAreaNumber = stopAreaString.match(/^\d+/)?.[0];
-    if (stopAreaNumber === undefined) continue;
-
-    const stopAreaName = query[stopAreaNumber];
-    if (!stopAreaName) continue;
-
-    const stopArea = selectedStops.value.find((s) => s.name === stopAreaName);
-    if (!stopArea) continue;
-
-    const stopPoints = stopAreaString.split("-");
-
-    for (let i = 1; i < stopPoints.length; i++) {
-      const stopPoint = stopArea.details.stopPoints.find((s) => s.id.endsWith(stopPoints[i]));
-      if (!stopPoint) continue;
-
-      excludedStopPoints.push([stopArea.id, stopPoint.id]);
-    }
-  }
-
-  return excludedStopPoints;
-}
-
-function serializeExcludedStopPoints(excludedStopPoints: [StopArea["id"], StopPoint["id"]][]) {
-  const partialExcludedStopPoints = [] as string[];
-
-  for (const stopAreaId of excludedStopPoints.map(([stopArea]) => stopArea).filter(unique)) {
-    const stopArea = selectedStops.value.find((s) => s.id === stopAreaId);
-    if (!stopArea) continue;
-
-    const stopAreaNumber = Object.keys(query).find((k) => query[k] === stopArea.name);
-    if (stopAreaNumber === undefined) continue;
-
-    partialExcludedStopPoints.push(
-      `${stopAreaNumber}-${excludedStopPoints
-        .filter(([stopArea]) => stopArea === stopAreaId)
-        .map((esp) => esp[1].substring("stop_point:".length))
-        .join("-")}`,
-    );
-  }
-
-  return partialExcludedStopPoints.join(",");
-}
-
-function getWantedStops(stops: typeof selectedStops.value) {
-  return stops.reduce(
-    (acc, val) => [
-      ...acc,
-      ...val.details.stopPoints.filter(
-        (sp) =>
-          !excludedStopPoints.value.find(
-            ([stopAreaId, stopPointId]) => val.id === stopAreaId && sp.id === stopPointId,
-          ),
-      ),
-    ],
-    [] as StopPoint[],
-  );
-}
-
-function updateStoredSettings() {
-  setTimeout(() => {
-    Preferences.set({ key: preferencesKeys.settings, value: JSON.stringify(settings.value) });
-  }, 50);
+  return r >= 0;
 }
 </script>
